Change computation weights
Before: The weights given in contact_matrix(weights) gave the value of each individual
Now: The weights are the weight applied to the group.

Example with weekday:
Before:  participants who complete diaries for weekdays get a weight 5 and those who complete diaries for weekends get a weight 2
Now: The sum of the weights of the participants who complete diaries for weekdays is 5, the sum of the weights on weekend is 2. The individual weight is 5/N_weekday or 2/N_weekend.